Set non-generic list value to default value when Ion list has non-matching types
This PR supports permissive mode for the following scenario:

The serialized data is an Ion list containing different types (e.g. 1, "two", 3.14d) and we want to deserialize to a non-generic list.

If permissive mode is enabled, we will set the default value of the list type within the list.