Remove permissive check when creating List serializer
Originally I had a permissive check when creating a list serializer and if permissive mode was on, it would create a default list serializer. Subsequently it would iterate through the Ion list and add the values into the list.

This commit is to fail fast and the permissive check at the casting will return the default value of the type.